🎾add template layout test
