+++ META-INF/MANIFEST.MF
@@ -1,1 +1,1 @@
-Main-Class: com.fedapp.App
+Main-Class: com.elune.App
