+++ src/main/java/com/elune/init/AppLoader.java
@@ -1,5 +1,5 @@
 /**
- * Elune - Lightweight Forum Powered by Razor
+ * Elune - Lightweight Forum Powered by Razor.
  * Copyright (C) 2017, Touchumind<chinash2010@gmail.com>
  * <p>
  * This program is free software: you can redistribute it and/or modify
@@ -16,42 +16,25 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.fedapp.controller;
 
-import com.fedapp.App;
-import com.fedepot.ioc.annotation.FromService;
-import com.fedepot.mvc.annotation.Route;
-import com.fedepot.mvc.controller.Controller;
+package com.elune.init;
 
-import java.util.HashMap;
-import java.util.Map;
-
-import static com.fedapp.Constant.*;
+import com.elune.App;
+import com.fedepot.Razor;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
 
 /**
- * Home controller
+ * Application loader
  *
  * @author Touchumind
  * @since 0.0.1
  */
-public class HomeController extends Controller {
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public final class AppLoader {
 
-    @FromService
-    private App app;
+    public static void init(Razor razor, App app) {
 
-    @Route("")
-    public void index() {
-
-        String name = "Elune";
-
-        String des = "Configuration factory Ok";
-
-        Map<String, Object> model = new HashMap<>();
-
-        model.put("name", name);
-        model.put("des", des);
-        model.put("other", app.configuration().get(CONFIG_KEY_MYSQL_PASS, ""));
-
-        Render("index.htm", model);
+        // TODO
     }
 }
