+++ src/main/resources/WEB-INF/templates/404.htm
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Not Found</title>
+    <style>
+        h2 {
+            text-align: center;
+            border-bottom: 1px solid #eee;
+            padding: 20px;
+        }
+        p {
+            text-align: center;
+        }
+    </style>
+</head>
+<body>
+<h2>404 Not Found</h2>
+<p>Elune/Razor</p>
+</body>
+</html>
