+++ src/main/resources/WEB-INF/templates/inc/header.htm
@@ -1,0 +1,3 @@
+<header>
+    header
+</header>
