+++ src/main/resources/WWW/assets/css/app.css
@@ -1,0 +1,3 @@
+body {
+    background: #f5f5f5;
+}
