+++ src/main/resources/WWW/assets/js/app.js
@@ -1,0 +1,1 @@
+var app = "elune";
