+++ src/main/resources/WWW/assets/js/vender.js
@@ -1,0 +1,1 @@
+var vender = true;
